Rename package sql/tpp into sql/parser
